personal_finance_manager_web/requirements.txt
# requirements.txt
Flask
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy # <-- Add this line!
python-dotenv
Flask-Login
Flask-Bcrypt
matplotlib